cxl-graph: Add cxl graph command to construct CXL topology graph images
This adds a new command "cxl graph" to plot the cxl topology graph or
dump cxl topology to a json formatted file. The command requires cxl
port/memdev has a contribute named "parent_dport" which is leveraged to
generate the topology.
The CXL topology can be either read from the file identified from
"--input" option or parsed at runtime as "cxl list" does, and output
will be stored in a file given by "--output".

This addition makes it easier to visualize the CXL topology.
Also, being able to choose to use an input text file
allows one to visualize the CXL topology while not having
access to CXL hardware, which may be convenient.

Signed-off-by: Fan Ni <fan.ni@samsung.com>

The change is based on Matthew Ho' patch as list below[1].
The main changes include,
    1. Remove the `root port` attribute for plotting the topology,
    instead leveraging the `parent_dport` attribute of the memdev and port
    objects to plot the topology.
    2. Fix the error messages, using error() instead of printf.
    3. Move all the graph related the functions to graph.c.
    4. add a new '-t' option indicating the output format (graph or json
    formatted plain file).
    5. Instead of implementing the topology plotting function as a
    subcommand of 'cxl list', we add a new cxl command 'cxl graph', and use
    output file type (indicated by the suffix of the file identified by
    the `-o` option) to determine plotting the graph or dump to json
    formatted plain file.
    6. Add Documentation/cxl/cxl-graph.txt
    7. Fix some bugs in graph plotting related functions.

[1] Matthew Ho's patch:
https://lore.kernel.org/linux-cxl/cover.1660895649.git.sunfishho12@gmail.com/